# ih-backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "iHoliday API",
        "description": "Base API documentation for the iHoliday services.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/api/v1/hotel/search": {
            "post": {
                "tags": [
                    "Hotels"
                ],
                "summary": "Search hotels for a city and dates",
                "operationId": "hotelSearch",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "city",
                                    "checkIn",
                                    "checkOut"
                                ],
                                "properties": {
                                    "city": {
                                        "type": "string",
                                        "example": "Goa"
                                    },
                                    "checkIn": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-11-10"
                                    },
                                    "checkOut": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-11-15"
                                    },
                                    "rooms": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "adults": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Hotel search results"
                    }
                }
            }
        },
...
        "/api/v1/hotel/book": {
            "post": {
                "tags": [
                    "Hotels"
                ],
                "summary": "Create hotel booking",
                "operationId": "hotelBook",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "hotel",
                                    "guests"
                                ],
                                "properties": {
                                    "hotel": {
                                        "type": "object"
                                    },
                                    "guests": {
                                        "type": "array",
                                        "items": {
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Hotel booking created"
                    }
                }
            }
        },
...
    },
...
}